Fix overflow in performance calculation

+++ FFT/src/host/fft_benchmark.cpp
@@ -96,7 +96,7 @@ fft::FFTBenchmark::executeKernel(FFTData &data) {
 
 void
 fft::FFTBenchmark::printResults(const fft::FFTExecutionTimings &output) {
-    double gflop = 5 * (1 << LOG_FFT_SIZE) * LOG_FFT_SIZE * executionSettings->programSettings->iterations * 1.0e-9;
+    double gflop = static_cast<double>(5 * (1 << LOG_FFT_SIZE) * LOG_FFT_SIZE) * executionSettings->programSettings->iterations * 1.0e-9;
 
     double minTime = *min_element(output.timings.begin(), output.timings.end());
     double avgTime = accumulate(output.timings.begin(), output.timings.end(), 0.0)
